refactor: group worker trace helpers
